+++ cognite/neat/_graph/transformers/_rdfpath.py
@@ -1,63 +1,12 @@
 from typing import cast
 from urllib.parse import quote
 
-from rdflib import Graph, Namespace, URIRef
+from rdflib import Namespace, URIRef
 from rdflib.query import ResultRow
 
-from cognite.neat._rules.models._rdfpath import RDFPath, SelfReferenceProperty, SingleProperty
-from cognite.neat._rules.models.information import InformationRules
 from cognite.neat._utils.rdf_ import get_namespace, remove_namespace_from_uri
 
-from ._base import BaseTransformer, BaseTransformerStandardised, RowTransformationOutput
-
-
-class ReduceHopTraversal(BaseTransformer):
-    """ReduceHopTraversal is a transformer that reduces the number of hops to direct connection."""
-
-    ...
-
-
-# TODO: Standardise
-class AddSelfReferenceProperty(BaseTransformer):
-    description: str = "Adds property that contains id of reference to all references of given class in Rules"
-    _use_only_once: bool = True
-    _need_changes = frozenset({})
-    _ref_template: str = """SELECT ?s WHERE {{?s a <{type_}>}}"""
-
-    def __init__(
-        self,
-        rules: InformationRules,
-    ):
-        self.rules = rules
-        self.properties = [
-            prop_
-            for prop_ in rules.properties
-            if prop_.instance_source and isinstance(prop_.instance_source.traversal, SelfReferenceProperty)
-        ]
-
-    def transform(self, graph: Graph) -> None:
-        for property_ in self.properties:
-            # Checked the instance source in the init.
-            prefix = property_.instance_source.traversal.class_.prefix  # type: ignore [union-attr]
-            suffix = property_.instance_source.traversal.class_.suffix  # type: ignore [union-attr]
-
-            namespace = self.rules.prefixes[prefix] if prefix in self.rules.prefixes else self.rules.metadata.namespace
-
-            for (reference,) in graph.query(self._ref_template.format(type_=namespace[suffix])):  # type: ignore [misc]
-                graph.add(
-                    (
-                        reference,
-                        self.rules.metadata.namespace[property_.property_],
-                        reference,
-                    )
-                )
-
-            traversal = SingleProperty.from_string(
-                class_=str(property_.class_),
-                property_=f"{self.rules.metadata.prefix}:{property_.property_}",
-            )
-
-            property_.instance_source = RDFPath(traversal=traversal)
+from ._base import BaseTransformerStandardised, RowTransformationOutput
 
 
 class MakeConnectionOnExactMatch(BaseTransformerStandardised):